fix(reporting): address review feedback
Move dry-run planning to a composed evaluator, remove unimplemented backend exports, skip event allocation without a reporter, and document the reporting API.
