+++ ccflow/evaluators/__init__.py
@@ -1,3 +1,4 @@
 from .common import *
+from .dry_run import *
 from .reporting import *
 from .retry import *
