+++ docs/wiki/reference/Reference.md
@@ -6,6 +6,7 @@ Technical description of the pieces `ccflow` provides. These pages are for looki
 - **[Contexts and Results](Contexts-and-Results)** — the built-in context and result types.
 - **[Built-in Models](Built-in-Models)** — the models, publishers, and evaluators that ship with `ccflow`, plus cache-key derivation.
 - **[Flow Model](Flow-Model)** — the `@Flow.model` API for defining callable models from plain Python functions.
+- **[Reporting](Reporting)** — lifecycle events, reporter sinks, reporting evaluators and models, dry-run planning, and OpenTelemetry integration.
 
 > [!NOTE]
 > Some items are marked *Coming Soon!* — they are part of the design and are in the process of being open-sourced.
